system_design/src/auth/requirements.txt
astroid==3.0.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.7
Flask==3.0.0
Flask-MySQLdb==2.0.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mysqlclient==2.2.0
parso==0.8.3
platformdirs==3.11.0
PyJWT==2.8.0
pylint==3.0.2
requests==2.31.0
tomlkit==0.12.2
urllib3==2.0.7
Werkzeug==3.0.1
